fMongo/flask1_mongo/requirements.txt
flask
flask_restful
flask_mongoengine
PyJWT
pymysql
apscheduler
flask.ext